// src/app/service/user/user-response-Interface.ts
export interface User {
  id?: number | string;
  username: string | null;
  gender: string | null;
  range: string | null;
  bio: string | null;
  designation: string | null;
  package: string | null;
  role: string | null;
  expertise: string | null;
  degree: string | null;
  mobile: string | null;
}